fixed deprecations

+++ utils/loss.py
@@ -225,7 +225,9 @@ class ComputeLoss:
             gij = (gxy - offsets).long()
             gi, gj = gij.T  # grid indices
 
-            # Append
+            # Append            
+            #print(f"gain: {gain}")
+            gain=gain.long()  #  added by anguelos
             indices.append((b, a, gj.clamp_(0, gain[3] - 1), gi.clamp_(0, gain[2] - 1)))  # image, anchor, grid indices
             tbox.append(torch.cat((gxy - gij, gwh), 1))  # box
             anch.append(anchors[a])  # anchors
